migrated data page components
