+++ sites/geohub/src/components/pages/data/datasets/PublishedDatasetDeleteDialog.svelte
@@ -1,20 +1,22 @@
 <script lang="ts">
 	import { ModalTemplate, Notification } from '@undp-data/svelte-undp-components';
-	import { createEventDispatcher } from 'svelte';
 
-	const dispatch = createEventDispatcher();
-
-	export let id: string;
-	export let name: string;
-
-	export let dialogShown = false;
-
-	$: if (dialogShown) {
-		openDeleteDialog();
+	interface Props {
+		id: string;
+		name: string;
+		dialogShown?: boolean;
+		ondelete?: () => void;
 	}
 
-	let deletedDatasetName = '';
-	let isDeleting = false;
+	let {
+		id = $bindable(),
+		name = $bindable(),
+		dialogShown = $bindable(false),
+		ondelete = () => {}
+	}: Props = $props();
+
+	let deletedDatasetName = $state('');
+	let isDeleting = $state(false);
 
 	const handleDeleteDataset = async () => {
 		try {
@@ -24,10 +26,9 @@
 				method: 'DELETE'
 			});
 			if (res.ok && res.status === 204) {
-				dispatch('deleted', {
-					id,
-					name
-				});
+				if (ondelete) {
+					ondelete();
+				}
 				closeDeleteDialog();
 			}
 		} finally {
@@ -56,34 +57,45 @@
 		const root = document.documentElement;
 		root.classList.remove('is-clipped');
 	};
+	$effect(() => {
+		if (dialogShown === true) {
+			openDeleteDialog();
+		}
+	});
 </script>
 
 {#if dialogShown}
 	<ModalTemplate title="Are you sure deleting the dataset?" bind:show={dialogShown}>
-		<div slot="content">
-			<div class="is-size-6 has-text-weight-normal">
-				<Notification type="warning" showCloseButton={false}>
-					Unexpected bad things will happen if you don't read this!
-				</Notification>
-				<div class="has-text-weight-medium mt-2">
-					This action <b>cannot</b> be undone. This will delete
-					<b>{name}</b>
-					from GeoHub data catalogue. It will not be searchable again from Data tab in GeoHub app.
+		{#snippet content()}
+			<div>
+				<div class="is-size-6 has-text-weight-normal">
+					<Notification type="warning" showCloseButton={false}>
+						Unexpected bad things will happen if you don't read this!
+					</Notification>
+					<div class="has-text-weight-medium mt-2">
+						This action <b>cannot</b> be undone. This will delete
+						<b>{name}</b>
+						from GeoHub data catalogue. It will not be searchable again from Data tab in GeoHub app.
+						<br />
+						Please type <b>{name}</b> to confirm.
+					</div>
 					<br />
-					Please type <b>{name}</b> to confirm.
+					<input class="input" type="text" bind:value={deletedDatasetName} />
 				</div>
-				<br />
-				<input class="input" type="text" bind:value={deletedDatasetName} />
 			</div>
-		</div>
-		<div slot="buttons">
-			<button
-				class="button is-primary {isDeleting ? 'is-loading' : ''} is-uppercase has-text-weight-bold"
-				on:click={handleDeleteDataset}
-				disabled={deletedDatasetName !== name}
-			>
-				delete this dataset
-			</button>
-		</div>
+		{/snippet}
+		{#snippet buttons()}
+			<div>
+				<button
+					class="button is-primary {isDeleting
+						? 'is-loading'
+						: ''} is-uppercase has-text-weight-bold"
+					onclick={handleDeleteDataset}
+					disabled={deletedDatasetName !== name}
+				>
+					delete this dataset
+				</button>
+			</div>
+		{/snippet}
 	</ModalTemplate>
 {/if}
